chore: update urls
