+++ bd_api/apps/account_auth/urls.py
@@ -4,7 +4,7 @@ from django.urls import path
 from . import views
 
 urlpatterns = [
-    path("", views.auth, name="auth"),
-    path("login/", views.signin, name="login"),
-    path("logout/", views.signout, name="logout"),
+    path("auth/", views.auth, name="auth"),
+    path("auth/login/", views.signin, name="login"),
+    path("auth/logout/", views.signout, name="logout"),
 ]
