+++ bd_api/apps/account_payment/urls.py
@@ -1,4 +1,4 @@
 # -*- coding: utf-8 -*-
 from django.urls import include, path
 
-urlpatterns = [path("", include("djstripe.urls", namespace="payment"))]
+urlpatterns = [path("payment/", include("djstripe.urls", namespace="payment"))]
