+++ bd_api/urls.py
@@ -18,30 +18,13 @@ from django.conf import settings
 from django.conf.urls.static import static
 from django.contrib import admin
 from django.urls import include, path
-from django.views.decorators.csrf import csrf_exempt
-from graphene_file_upload.django import FileUploadGraphQLView
-
-from bd_api.apps.api.v1.search_views import DatasetSearchView as DatasetSearchV2View
-from bd_api.apps.api.v1.views import DatasetRedirectView
-from bd_api.apps.api.v1.views import DatasetSearchView as DatasetSearchV1View
-
-
-def graphql_view():
-    return csrf_exempt(FileUploadGraphQLView.as_view(graphiql=True))
-
 
 urlpatterns = [
     path("admin/", admin.site.urls),
     path("", include("bd_api.apps.core.urls")),
-    path("api/", include("bd_api.apps.api.v1.urls")),
-    path("api/graphql/", graphql_view()),
-    path("account/", include("bd_api.apps.account.urls")),
-    path("auth/", include("bd_api.apps.account_auth.urls")),
-    path("search/", DatasetSearchV1View.as_view()),
-    path("search/v2/", DatasetSearchV2View.as_view()),
-    path("search/debug/", include("haystack.urls")),
-    path("dataset/", DatasetRedirectView.as_view()),
-    path("dataset_redirect/", DatasetRedirectView.as_view()),
-    path("payment/", include("bd_api.apps.account_payment.urls")),
+    path("", include("bd_api.apps.api.v1.urls")),
+    path("", include("bd_api.apps.account.urls")),
+    path("", include("bd_api.apps.account_auth.urls")),
+    path("", include("bd_api.apps.account_payment.urls")),
 ]
 urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
